backend/requirements.txt
Flask>=2.0
flask-cors>=3.0.0
openai>=0.27.0
python-dotenv>=0.19.0